+++ iOSChallenge/iOSChallenge/Models/Common/ParkingSpace.swift
@@ -1,0 +1,11 @@
+//
+//  ParkingSpace.swift
+//  iOSChallenge
+//
+//  Created by Jacobo Adrián Sande Veiga on 16/5/25.
+//
+
+struct ParkingSpace: Codable {
+    let hasParkingSpace: Bool
+    let isParkingSpaceIncludedInPrice: Bool
+}
